requirements.txt
asgiref==3.4.1
cycler==0.10.0
Django==3.2.6
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
opencv-python==4.5.3.56
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
